+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ### Fixes
 
+* Fixed time series legend entry to show the dataset title associated with 
+  the time series. (#574)
+
 * Show dataset level in the zoom level info box also if no dataset variable is being 
   displayed (when in RGB-only mode). (#576)